Backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
gspread
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2

